# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartStock inventory and order management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/google:
    get:
      operationId: startGoogleAuth
      tags: [auth]
      summary: Start Google OAuth
      responses:
        "302":
          description: Redirects to Google when configured
        "503":
          description: Google OAuth is not configured
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Sign in to the demo admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Invalid credentials
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List and filter products
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
        - name: sort
          in: query
          schema:
            type: string
            enum: [price-asc, price-desc, newest]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
        "401":
          description: Authentication required
        "403":
          description: Admin role required
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Authentication required
        "403":
          description: Admin role required
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "401":
          description: Authentication required
        "403":
          description: Admin role required
        "404":
          description: Not found
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders visible to the current user
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Validate stock and place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input or insufficient stock
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order visible to the current user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update an order status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid transition
        "403":
          description: Admin role required
        "404":
          description: Not found
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "403":
          description: Admin role required
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "403":
          description: Admin role required
        "404":
          description: Not found
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "403":
          description: Admin role required
        "404":
          description: Not found
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory with status
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
        "403":
          description: Admin role required
  /inventory/{id}:
    put:
      operationId: adjustInventory
      tags: [inventory]
      summary: Adjust product inventory
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustmentInput"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "400":
          description: Invalid adjustment
        "403":
          description: Admin role required
        "404":
          description: Not found
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Search:
      name: search
      in: query
      schema:
        type: string
    Category:
      name: category
      in: query
      schema:
        type: string
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        profileImage: { type: ["string", "null"] }
        role:
          type: string
          enum: [customer, admin]
      required: [id, name, email, profileImage, role]
    Session:
      type: object
      properties:
        authenticated: { type: boolean }
        user:
          anyOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        googleConfigured: { type: boolean }
      required: [authenticated, user, googleConfigured]
    AdminLoginInput:
      type: object
      properties:
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 1 }
      required: [email, password]
    Supplier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        createdAt: { type: string }
      required: [id, name, contactPerson, email, phone, address, createdAt]
    SupplierInput:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        contactPerson: { type: string, minLength: 2 }
        email: { type: string, minLength: 5 }
        phone: { type: string, minLength: 5 }
        address: { type: string, minLength: 3 }
      required: [name, contactPerson, email, phone, address]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        quantity: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
        supplier:
          anyOf:
            - $ref: "#/components/schemas/Supplier"
            - type: "null"
        image: { type: string }
        stockStatus:
          type: string
          enum: [in-stock, low-stock, out-of-stock]
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, sku, category, description, price, quantity, minimumStock, supplier, image, stockStatus, createdAt, updatedAt]
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        sku: { type: string, minLength: 2 }
        category: { type: string, minLength: 2 }
        description: { type: string, minLength: 5 }
        price: { type: number, minimum: 0 }
        quantity: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
        supplierId: { type: ["string", "null"] }
        image: { type: string, minLength: 1 }
      required: [name, sku, category, description, price, quantity, minimumStock, image]
    OrderItem:
      type: object
      properties:
        productId: { type: string }
        productName: { type: string }
        price: { type: number }
        quantity: { type: number, minimum: 1 }
        subtotal: { type: number }
      required: [productId, productName, price, quantity, subtotal]
    Order:
      type: object
      properties:
        id: { type: string }
        customerName: { type: string }
        customerEmail: { type: string }
        customerPhone: { type: string }
        address: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        totalAmount: { type: number }
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, customerName, customerEmail, customerPhone, address, items, totalAmount, status, createdAt, updatedAt]
    OrderInput:
      type: object
      properties:
        customerName: { type: string, minLength: 2 }
        customerEmail: { type: string, minLength: 5 }
        customerPhone: { type: string, minLength: 7 }
        address: { type: string, minLength: 5 }
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              productId: { type: string }
              quantity: { type: number, minimum: 1 }
            required: [productId, quantity]
      required: [customerName, customerEmail, customerPhone, address, items]
    OrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
      required: [status]
    InventoryItem:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        currentStock: { type: number }
        minimumStock: { type: number }
        stockStatus:
          type: string
          enum: [in-stock, low-stock, out-of-stock]
      required: [product, currentStock, minimumStock, stockStatus]
    InventoryAdjustmentInput:
      type: object
      properties:
        quantity: { type: number }
        reason:
          type: string
          enum: [restock, manual-adjustment, damaged, returned]
      required: [quantity, reason]
...