# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SmartStock inventory and order management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  /** @minLength 2 */
  customerName: string;
  /** @minLength 5 */
  customerEmail: string;
  /** @minLength 7 */
  customerPhone: string;
  /** @minLength 5 */
  address: string;
  /** @minItems 1 */
  items: OrderInputItemsItem[];
}