# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SmartStock inventory and order management API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStockStatus } from './productStockStatus';
import type { Supplier } from './supplier';

export interface Product {
  id: string;
  name: string;
  sku: string;
  category: string;
  description: string;
  price: number;
  /** @minimum 0 */
  quantity: number;
  /** @minimum 0 */
  minimumStock: number;
  supplier: Supplier | null;
  image: string;
  stockStatus: ProductStockStatus;
  createdAt: string;
  updatedAt: string;
}